Fix values table serialization
